requirements.txt
pandas==2.1.0
psycopg2-binary
avro
pyarrow
sqlalchemy
xlrd
python-snappy
astronomer-cosmos[dbt-postgres]
apache-airflow-providers-telegram==4.1.1 